Implement Details Screen

+++ app/src/main/java/com/example/bmmovies/domain/entities/local/MovieDetails.kt
@@ -1,14 +1,14 @@
 package com.example.bmmovies.domain.entities.local
 
 data class MovieDetails(
-    val id:Long=0,
-    val title:String="",
-    val posterImage:String="",
-    val releaseDate:String="",
-    val rating:String="",
-    val overview:String="",
-    val backdropImage:String="",
-    val voteCount:String="",
-    val isAdult:String="",
-    val trailers:List<Trailer> = listOf()
+    val id: Int = 0,
+    val title: String = "",
+    val posterImage: String = "",
+    val releaseDate: String = "",
+    val rating: String = "",
+    val overview: String = "",
+    val backdropImage: String = "",
+    val voteCount: String = "",
+    val isAdult: String = "",
+    val trailers: List<Trailer> = listOf()
 )
