+++ app/src/main/java/com/example/bmmovies/domain/entities/query/MovieDetailsQuery.kt
@@ -2,8 +2,7 @@ package com.example.bmmovies.domain.entities.query
 
 
 data class MovieDetailsQuery(
-    var movieId: Long,
+    var movieId: Int,
     val language: String = "en-US",
     val appendToResponse: String = "videos",
-//    var apiKey: String = "95f7a492542e065f89ea88ff37aa8122"
 )
